core/standards/scripts/scheduler2/location_hash/310.html

<!doctype html>
<title>Location hash - set hash in external script with defer</title>
<script src="/resources/testharness.js"></script><script src="/resources/testharnessreport.js"></script>
<script>
var t = async_test();
var histlength;
t.step(function() {
    if( location.hash && location.hash != '#' ) {
        location.href = location.href.replace(/#.*$/,'');
        return;
    }
    histlength = history.length;
});
</script>
<script>
t.step(function() {
    location.hash = '#foo';
});
</script>
<script src="310.js" id="test_script" defer></script>
<script>
t.step(function() {
    //var s = document.getElementById("test_script");
    //assert_true( s.defer );
    assert_equals( location.hash, '#foo' );
    assert_equals( history.length, histlength );
});
t.done();
</script>
<div id="log"></div>
